pyproject.toml
[tool.poetry]
name = "KB_Builder"
version = "0.1.0"
description = ""
authors = ["ahmedmukhtar <amaamsmnbeai@gmail.com>"]
readme = "README.md"
packages = [{include = "KB_Builder"}]

[tool.poetry.dependencies]
python = "^3.9.13"
sentence-transformers = "^2.2.2"
chromadb = "^0.3.11"
moviepy = "^1.0.3"
ipykernel = "^6.21.3"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
numpy = "1.23.5"
ffmpeg-python = "^0.2.0"
black = "^23.1.0"
python-dotenv = "^1.0.0"
networkx = "^3.0"
pandas = "^1.5.3"
matplotlib = "^3.7.1"
plotly = "^5.13.1"
nbformat = "^5.7.3"
ffmpeg = "^1.4"
pinecone-client = "^2.2.4"
youtube-transcript-api = "^0.6.1"
openai = "^1.3.7"
langchain = "^0.0.346"
gradio = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
